Improve performance: just one regex for each module list
Rather than looping over all the module names on each line of each file,
build a single regex from the relevant module names, and test each line
against it.

In the worst case, this moves a loop from code written in Perl to the C
implementation of the Perl regexp engine. On Perl ≥ 5.10, the improvement
is even better, because the regexp engine has a better optimisation for
regexps with lots of fixed-text alternations.

On my machine, this reduces execution time of _strict_ok() on an older
version of lib/Test/Strict.pm from 602 µs ± 0.1% to 50 µs ± 0.2% (according
to Dumbbench).