Don't simultaneously test -c and -v switches
Up through perl-5.37.3, there was a bug in perl (see
https://github.com/Perl/perl5/issues/20252#issuecomment-1236880872)
which permitted simultaneous use of the '-c' and '-v' switches to the
perl interpreter.  Once this bug was corrected, t/01all.t began to fail.
As Test::Strict has many CPAN distributions depending on it, this test
failure has had wide impact.

This pull request modifies tests appropriately and has been tested on
bleadperl (v5.37.7-134-g25948dfb24) and on perl-5.32.1, in both cases on
threaded builds on FreeBSD-12.  Devel::Cover was not installed on
bleadperl, hence t/04cover.t was skipped on that build.  Otherwise, all
tests are passing.  In the course of testing, certain additions were
made to .gitignore and MANIFEST.SKIP to get metadata-oriented tests to
pass.

NOTE:  In my git checkout I find the following file:

        .git/hooksfsmonitor-watchman.sample

This is deemed by 'file' to be a Perl executable.  As such, it is pushed
onto the list of files tested by 'all_perl_files_ok()' at the start of
t/01all.t.  This in turn means that in my 'git checkout' the count of
tests is off by 2 and the test is reported to fail.

        "Looks like you planned 59 tests but ran 61."

Because the files in .git/hooks/ should not be included in the tarball,
the count of tests is correct when run from the tarball.  I leave to the
maintainer to determine how to square this circle.  You should probably
check for the presence of a .git/ directory and exclude that directory
from the list of directories passed to all_perl_files_ok().

For https://github.com/manwar/Test-Strict/issues/32